Redesign Guest Representation,
Redesigned guest representation, guest is not a state anymore.
instead, we have an interface IUser, which is implemented by RegisteredUser, and GuestUser. the RegisteredUser still has his states.
Getting the appropriate user is done using the GetUserObject function of the UserDomain.
According to the the new Guest design, I changed the Guid which has been the identifier of the user, to a UserIdentifier, which has a Guid like before, and a boolean IsGuest
Fixed bugs.
Added support for SearchProduct
Fixed AddOwner to not rely on AddManager
Fixed tests.
Changed tests to address the ServiceFacadeProxy instead, made things easier.
More refactoring to names and function signatures
Added clear method to SessionManager too
Disabled UnitTesting untill they are rewritten