Finished Server - Side Notifications Implementation
TODO:
Tests notifications after merge with AddingPresentationLayer
Support persistent notifications
Maybe create Initialize class to initialize system, and subscribe, etc.

Did:
Added an IUpdateEvent interface in ApplicationCore, this interface defines an update event.
Added a RemovedOwnerEvent class in ApplicationCore, which implements IUpdateEvent.
Added in DomainLayerFacade, at the end of CascadeRemoveOwner, check if the operation succeeded - create a RemovedOwnerEvent, and send it to the UpdateCenter
UpdateCenter is a static class in the DomainLayer, upon receiving an IUpdateEvent, it notifies all its subscribers (calls their subscribed delgate)
NotificationSender class subscribes to the UpdateCenter. when the UpdateCenter receives an IUpdateEvent, NotificationSender is notified and sends the notification to the StronglyTypedNotifcationHub, which now implements the IUserNotifier interface.
The Hub then sends the notification to the list of users defined.

Added a new ConnectionManager class to map between sessionId and the ConnectionId, using this class we know what is the connetionId of the user we want to send the notification to.