Moved Notifications saving to NotificationsCenter (previously Sender)
Also changed the flow so the Hub checks for every new connection if he should be notified.
This is also needed because we always redirect after an action, so the sending will be displayed for a split second -- or we will get an exception.
Should search online for persistent connection with SignalR